[refactor] Change navigation and styling code
This is a major refactor to remove large parts of the template app code and change the navigation between screens. TypeScript is causing me a few issues so refactored to scale back on its use for now. Moved some functionality between App.tsx and components to make it easier.
