+++ types.tsx
@@ -3,9 +3,8 @@
  * https://reactnavigation.org/docs/typescript/
  */
 
-import { BottomTabScreenProps } from '@react-navigation/bottom-tabs';
-import { CompositeScreenProps, NavigatorScreenParams } from '@react-navigation/native';
-import { NativeStackScreenProps } from '@react-navigation/native-stack';
+import { CompositeScreenProps, NavigatorScreenParams, RouteProp } from '@react-navigation/native';
+import { NativeStackScreenProps, NativeStackNavigationProp, createNativeStackNavigator } from '@react-navigation/native-stack';
 
 declare global {
   namespace ReactNavigation {
@@ -14,8 +13,8 @@ declare global {
 }
 
 export type RootStackParamList = {
-  Root: NavigatorScreenParams<RootTabParamList> | undefined;
-  Modal: undefined;
+  Home: undefined;
+  Forecast: undefined;
   NotFound: undefined;
 };
 
@@ -24,12 +23,11 @@ export type RootStackScreenProps<Screen extends keyof RootStackParamList> = Nati
   Screen
 >;
 
-export type RootTabParamList = {
-  TabOne: undefined;
-  TabTwo: undefined;
-};
+// export type HomeParamList = {
+//   Forecast: undefined;
+// }
 
-export type RootTabScreenProps<Screen extends keyof RootTabParamList> = CompositeScreenProps<
-  BottomTabScreenProps<RootTabParamList, Screen>,
-  NativeStackScreenProps<RootStackParamList>
->;
+// export type HomeScreenProps = NativeStackScreenProps<
+// HomeParamList,
+// 'Forecast'
+// >;
